requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
requests
python-dotenv
numpy
opencv-python-headless # Use headless if you don't need GUI components
pydicom
nibabel
# SimpleITK # Uncomment if you decide to use it for advanced DICOM/NIfTI handling
torch # Install from pytorch.org for specific CUDA version if needed
torchvision # Install with torch
torchxrayvision
TotalSegmentator
monai[all] # Installs MONAI with common dependencies including nibabel, Pillow, etc.
# accelerate # Often useful with Hugging Face Transformers if you add such models
